Add CI config and an example test
One advantage to building a static site with nodejs is that we can add
some simple tests that will help to verify the data we're trying to
input into the system. CI will check that every pull request into master
is passing some minimum criteria.

- Pull a default configuration from cirle ci
- Add one example of a test that checks all the files in our posts
directory and confirms that they have the front-matter we're looking for